+++ crates/rbx-env/src/commands/mod.rs
@@ -2,3 +2,4 @@ pub mod gen_module;
 pub mod get;
 pub mod list;
 pub mod rm;
+pub mod set;